+++ src/main/java/de/wwu/scdh/teilsp/services/extensions/ConfiguredPluginLoader.java
@@ -1,7 +1,5 @@
 package de.wwu.scdh.teilsp.services.extensions;
 
-import java.awt.Frame;
-import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -9,7 +7,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.ServiceLoader;
 import javax.xml.namespace.NamespaceContext;
-import javax.xml.transform.URIResolver;
 import javax.xml.xpath.XPath;
 import javax.xml.xpath.XPathConstants;
 import javax.xml.xpath.XPathExpressionException;
@@ -19,9 +16,7 @@ import org.slf4j.LoggerFactory;
 
 import org.w3c.dom.Document;
 import org.w3c.dom.NodeList;
-import org.xml.sax.EntityResolver;
 
-import de.wwu.scdh.teilsp.ui.ISelectionDialog;
 import de.wwu.scdh.teilsp.config.ArgumentsConditionsPair;
 import de.wwu.scdh.teilsp.config.ExtensionConfiguration;
 import de.wwu.scdh.teilsp.config.ExtensionConfigurationReader;
@@ -32,23 +27,40 @@ import de.wwu.scdh.teilsp.xpath.XPathUtil;
 
 
 /**
- * {@link SelectionDialogLoader} - a class for loading plugins
- * registered for {@link ISelectionDialog} through the service
- * provider interface (SPI).
+ * {@link ConfiguredPluginLoader} - a class for loading plugins
+ * registered for an interface extending {@link ConfigurablePlugin}
+ * through the service provider interface (SPI).
  */
-public class SelectionDialogLoader {
+public class ConfiguredPluginLoader <T extends ConfigurablePlugin> {
 
-    public static final String DEFAULT_PROVIDER = "de.wwu.scdh.teilsp.ui.EditableComboBoxSelectDialog";
+    private final Logger LOGGER = LoggerFactory.getLogger(ConfiguredPluginLoader.class);
+
+    protected final String defaultProvider;
+
+    protected final Class<T> pluginInterface;
 
     /**
-     * Returns all providers in the classpath.
+     * Make a new loader for configured plugins implementing an
+     * interface.
+     *
+     * @param pluginInterface  the interface for which plugins shall be loaded
+     * @param defaultPlugin  the dotted name of the default plugin class
      *
      */
-    public static List<ISelectionDialog> providers() {
-	List<ISelectionDialog> services = new ArrayList<>();
-	ServiceLoader<ISelectionDialog> loader = ServiceLoader.load(ISelectionDialog.class);
-	loader.forEach(selectionDialog -> {
-		services.add(selectionDialog);
+    public ConfiguredPluginLoader(Class<T> pluginInterface, String defaultPlugin) {
+	this.pluginInterface = pluginInterface;
+	this.defaultProvider = defaultPlugin;
+    }
+
+    /**
+     * Returns all plugins present in the classpath.
+     *
+     */
+    public List<T> providers() {
+	List<T> services = new ArrayList<T>();
+	ServiceLoader<T> loader = ServiceLoader.load(pluginInterface);
+	loader.forEach(labelledEntriesProvider -> {
+		services.add(labelledEntriesProvider);
 	});
 	return services;
     }
@@ -57,107 +69,88 @@ public class SelectionDialogLoader {
      * Get provider by name.
      *
      */
-    public static ISelectionDialog provider(String providerName)
+    public T provider(String providerName)
 	throws ProviderNotFoundException {
-	ServiceLoader<ISelectionDialog> loader = ServiceLoader.load(ISelectionDialog.class);
-	Iterator<ISelectionDialog> iter = loader .iterator();
+	ServiceLoader<T> loader = ServiceLoader.load(pluginInterface);
+	Iterator<T> iter = loader .iterator();
 	while (iter.hasNext()) {
-	    ISelectionDialog provider = iter.next();
+	    T provider = iter.next();
 	    if (providerName.equals(provider.getClass().getName())) {
 		return provider;
 	    }
 	}
-	throw new ProviderNotFoundException("ISelectionDialog " + providerName + " not found");
+	throw new ProviderNotFoundException("ConfigurablePlugin " + pluginInterface.toString()
+					    + " " + providerName + " not found");
     }
 
     /**
      * Overloaded method, that returns the default provider.
      *
      */
-    public static ISelectionDialog provider()
+    public T provider()
 	throws ProviderNotFoundException {
-	return provider(DEFAULT_PROVIDER);
+	return provider(defaultProvider);
     }
 
     /**
      * This reads the configuration file and returns a list of
-     * initialized providers, defined for the current context.
+     * initialized plugins, defined for the current context.
      * @param document the current document
-     * @param systemId the url string pointing to the current document
      * @param context the current cursor position in the document
      * given by an XPath expression
      * @param nodeType the type of node which to offer completion suggestions
      * @param nodeName the name of the element or attribute node, which to offer suggestions for
-     * @param uriResolver a {@link URIResolver} to be used by the plugins
-     * @param entityResolver a {@link EntityResolver} to be used by the plugins
      * @param namespaceDecl the {@link NamespaceContext} of the current document
      * @param configFile an URL pointing to the config file, use
      * {@code file:...} for local files
-     * @param frame an Swing Frame object
      * @param expander an {@link EditorVariablesExpander}
      */
-    public static List<ISelectionDialog> providersForContext
+    public List<T> providersForContext
 	(Document document,
-	 String systemId,
 	 String context,
 	 String nodeType,
 	 String nodeName,
-	 URIResolver uriResolver,
-	 EntityResolver entityResolver,
 	 NamespaceContext namespaceDecl,
 	 String configFile,
-	 Frame frame,
 	 EditorVariablesExpander expander)
 	throws ExtensionException, ConfigurationException {
 	// read the plugin configuration from the config file
-	List<ExtensionConfiguration> extensionsConfiguration = ExtensionConfigurationReader.getExtensionsConfiguration(configFile, expander);
-	return providersForContext(document, systemId, context, nodeType, nodeName,
-				   uriResolver, entityResolver,
-				   namespaceDecl, frame, extensionsConfiguration);
+	List<ExtensionConfiguration> extensionsConfiguration =
+	    ExtensionConfigurationReader.getExtensionsConfiguration(configFile, expander);
+	return providersForContext(document, context, nodeType, nodeName,
+				   namespaceDecl, extensionsConfiguration);
     }
 
     /**
      * This returns a list of initialized providers, defined for the
      * current context.
      * @param document the current document
-     * @param systemId the url string pointing to the current document
      * @param context the current cursor position in the document
      * given by an XPath expression
      * @param nodeType the type of node which to offer completion suggestions
      * @param nodeName the name of the element or attribute node, which to offer suggestions for
-     * @param uriResolver a {@link URIResolver} to be used by the plugins
-     * @param entityResolver a {@link EntityResolver} to be used by the plugins
      * @param namespaceDecl the {@link NamespaceContext} of the current document
-     * @param frame an Swing Frame object
      * @param extensionsConfiguration the parsed config file
      * {@code file:...} for local files
      */
-    public static List<ISelectionDialog> providersForContext
+    public List<T> providersForContext
 	(Document document,
-	 String systemId,
 	 String context,
 	 String nodeType,
 	 String nodeName,
-	 URIResolver uriResolver,
-	 EntityResolver entityResolver,
 	 NamespaceContext namespaceDecl,
-	 Frame frame,
 	 List<ExtensionConfiguration> extensionsConfiguration)
 	throws ExtensionException, ConfigurationException {
+	List<T> providers = new ArrayList<T>();
 
 	// setup logging utility
-	final Logger LOGGER = LoggerFactory.getLogger(SelectionDialogLoader.class);
 
 	LOGGER.debug("Searching plugins for context: \"{}\"", context);
 	LOGGER.debug("Searching pluging for nodeType: {}, nodeName: {}", nodeType, nodeName);
 	LOGGER.debug("Implementation of document: {}", document.getClass().getCanonicalName());
 
-	// empty list, to be filled and returned
-	List<ISelectionDialog> providers = new ArrayList<ISelectionDialog>();
-
 	// Load plugins
-	// this is not providers, because they are from no-argument constructor
-	List<ISelectionDialog> plugins = SelectionDialogLoader.providers();
+	List<T> plugins = providers();
 
 	// we need some iteration variables
 	int i, j, k;
@@ -174,7 +167,7 @@ public class SelectionDialogLoader {
 	    config = extensionsConfiguration.get(i);
 	    // iteratre over loaded plugins
 	    for (j = 0; j < plugins.size(); j++) {
-		ISelectionDialog plugin = plugins.get(j);
+		T plugin = plugins.get(j);
 		// check if the configuration is for this provider
 		if (plugin.getClass().getCanonicalName().equals(config.getClassName())) {
 		    // check all specifications of this provider
@@ -205,10 +198,7 @@ public class SelectionDialogLoader {
 				// we make a new instance of the
 				// provider, because we do not want to
 				// configure the same several times.
-				//ISelectionDialog p = plugin.getClass().newInstance();
-				// this way we can also call an arbitrary constructor
-				ISelectionDialog p =
-				    (ISelectionDialog) plugin.getClass().getDeclaredConstructor(Frame.class).newInstance(frame);
+				T p = (T) plugin.getClass().newInstance();
 				p.init(arguments);
 				providers.add(p);
 			    }
@@ -221,16 +211,6 @@ public class SelectionDialogLoader {
 							     + contextXPath);
 			} catch (NullPointerException e) {
 			    throw new ConfigurationException("Configuration error\n" + e);
-			} catch (NoSuchMethodException e) {
-			    throw new ConfigurationException("Error loading dialog class "
-							     + plugin.getClass()
-							     + "\n" + "no constructor signature (java.awt.Frame frame)"
-							     + "\n" + e);
-			} catch (InvocationTargetException e) {
-			    throw new ConfigurationException("Error loading dialog class "
-							     + plugin.getClass()
-							     + "\n" + "failed constructor call"
-							     + "\n" + e);
 			} catch (InstantiationException e) {
 			    throw new ConfigurationException("Error loading plugin "
 							     + plugin.getClass().getCanonicalName()
@@ -244,12 +224,11 @@ public class SelectionDialogLoader {
 		}
 	    }
 	}
-	// TODO: should we dispose all plugins or does GC the job?
-
 	LOGGER.debug("Configured plugins: {}", providers.size());
-	for (ISelectionDialog p : providers) {
+	for (ConfigurablePlugin p : providers) {
 	    LOGGER.debug("Plugin {} configured with arguments:\n{}",
-			 p.toString());
+			 p.toString(),
+			 p.getArguments().toString());
 	}
 	return providers;
     }