+++ src/main/java/de/wwu/scdh/teilsp/services/extensions/ILabelledEntriesProvider.java
@@ -1,51 +1,56 @@
 package de.wwu.scdh.teilsp.services.extensions;
 
 import java.util.List;
-import java.util.Map;
 import javax.xml.transform.URIResolver;
 
 import org.xml.sax.EntityResolver;
 import org.w3c.dom.Document;
 
+import de.wwu.scdh.teilsp.exceptions.ConfigurationException;
+
 
 /**
- * {@link ILabelledEntriesProvider} - an interface for plugins, that
+ * {@link ILabelledEntriesProvider} - an interface for plugins, which
  * provide labelled entries.
  *
+ * By separting the initialization/setup phase and the getting phase
+ * which can send user input to the plugin, this should be sufficient
+ * for plugins that need on-going user interaction to get labelled
+ * entries from a huge collection of entries etc.
+ *
+ * Note, that credentials can be passed in as arguments in the
+ * initialization phase,
+ * see {@link ConfigurablePlugin#init(java.util.Map)}.
  */
-public interface ILabelledEntriesProvider {
+public interface ILabelledEntriesProvider extends ConfigurablePlugin {
 
     /**
-     * This method initializes the provider with a setup from
-     * configuration and the current editing context. It is once
-     * called during the initialization phase.
+     * {@link setup} passes a setup of document and resolvers to the
+     * plugin. This method should be called once called during the
+     * initialization phase,
+     * after {@link ConfigurablePlugin#init(java.util.Map)}.
      *
-     * @param  kwargs      key value pairs with configuration parameters
      * @param  uriResolver URI resolver used in the editor
      * @param  entityResolver entity resolver used in the editor
      * @param  document    the currently edited file as DOM object
      * @param  systemId    URL of the current document
-     * @return             an ordered sequence of labelled entries
+     * @param  context     XPath to the current editing position
      */
-    public void init(Map<String, String> kwargs,
-		     URIResolver uriResolver,
-		     EntityResolver entityResolver,
-		     Document document,
-		     String systemId) throws ExtensionException;
-    
+    public void setup
+	(URIResolver uriResolver,
+	 EntityResolver entityResolver,
+	 Document document,
+	 String systemId,
+	 String context)
+	throws ExtensionException;
+
     /**
-     * This method must return an ordered collection of labelled
+     * {@link getLabelledEntries} returns an ordered collection of labelled
      * entries. It takes a map of key value pairs as arguments.
      *
-     * @param  userInput   what the user has typed so far	
+     * @param  userInput   what the user has typed so far
      * @return             an ordered sequence of labelled entries
      */
     public List<LabelledEntry> getLabelledEntries(String userInput) throws ExtensionException;
 
-    /**
-     * Return arguments. This might be informative for debugging and
-     * error messages.
-     */
-    public Map<String, String> getArguments();
-    
 }